Updated API, fixed documentation
